rename device_ids module to connected_cars module
