+++ server/__main__.py
@@ -5,7 +5,7 @@ import logging
 from apscheduler.schedulers.background import BackgroundScheduler  # type: ignore
 
 from database.database_controller import remove_old_messages, set_message_retention_period  # type: ignore
-from database.device_ids import clear_connected_cars  # type: ignore
+from server.database.connected_cars import clear_connected_cars  # type: ignore
 from database.connection import set_db_connection  # type: ignore
 from database.time import timestamp  # type: ignore
 from fleetv2_http_api.__main__ import main as run_server  # type: ignore
